Update SConscript to register sandesh-generated skeletons with SCons.
analytics_request_skeleton.cc is a build dependency of
contrail-query-engine, and so has to be registered with SCons. Call a
variant of SandeshGenCpp that registers those source files.

However, contrail-collector is already linking against another source
file (XXX: check which one) with the same symbols and build it
with analytics_request_skeleton.cc fails, so we filter them from vizd
and buildinfo.cc targets.

Change-Id: I97bac46e88b3a60bc5c396913df5f287d894c031